// src/styles/mixins.types.ts
import {FlattenSimpleInterpolation} from "styled-components";

export interface IDevice {
    tabletS: FlattenSimpleInterpolation;
    tabletM: FlattenSimpleInterpolation;
    tabletL: FlattenSimpleInterpolation;
    laptopS: FlattenSimpleInterpolation;
    laptopM: FlattenSimpleInterpolation;
    laptopL: FlattenSimpleInterpolation;
    desktopS: FlattenSimpleInterpolation;
    desktopM: FlattenSimpleInterpolation;
    desktopL: FlattenSimpleInterpolation;
}